(PC-9271) App native - Sécurité - Remplacer les modales de vérification du numéro de téléphone par des BottomContentPage - intégration
